Cleaned up rw.py to conform to pep8-compliant variable names
